fix slate circular dependencies
